Semana 05 videos

+++ app/src/main/java/com/lugares/model/Lugar.kt
@@ -24,7 +24,7 @@ data class Lugar(
     @ColumnInfo(name="longitud")
     val longitud: Double,
     @ColumnInfo(name="altura")
-    val altura: Int,
+    val altura: Double,
     @ColumnInfo(name="rutaAudio")
     val rutaAudio: String?,
     @ColumnInfo(name="rutaImagen")
